requirements.txt
azure-functions
mpmath==1.3.0
azure-identity
azure-keyvault-secrets
azure-storage-blob
requests
